+++ src/errors/validation-error.ts
@@ -1,0 +1,5 @@
+'use strict'
+
+export class ValidationError extends Error {
+
+}